refactor: domain 단 킆래스 명명 변경
