+++ src/main/kotlin/domain/Car.kt
@@ -1,0 +1,3 @@
+package domain
+
+data class Car(var carId: Int, var count: Int = 0)
