--- src/main/kotlin/domain/CarMovementCount.kt
@@ -1,3 +1,0 @@
-package step3.domain
-
-data class CarMovementCount(var carId: Int, var count: Int = 0)
